GONRG-1537 Finalize Partition Service
-SonarLint comments are fixed.

+++ provider/partition-gcp/src/main/java/org/opengroup/osdu/partition/provider/gcp/service/PartitionServiceImpl.java
@@ -98,10 +98,10 @@ public class PartitionServiceImpl implements IPartitionService {
   @Transactional
   @Override
   public PartitionInfo updatePartition(String partitionId, PartitionInfo partitionInfo) {
-    this.partitionPropertyEntityRepository.findByPartitionId(partitionId)
-        .orElseThrow(
-            () -> new AppException(HttpStatus.SC_NOT_FOUND, UNKNOWN_ERROR_REASON,
-                "An attempt to update not existing partition."));
+    if (!this.partitionPropertyEntityRepository.findByPartitionId(partitionId).isPresent()) {
+      throw new AppException(HttpStatus.SC_NOT_FOUND, UNKNOWN_ERROR_REASON,
+          "An attempt to update not existing partition.");
+    }
     List<PartitionPropertyEntity> partitionProperties = new ArrayList<>();
     for (Map.Entry<String, Property> entry : partitionInfo.getProperties().entrySet()) {
       PartitionPropertyEntity entity = this.partitionPropertyEntityRepository
@@ -144,10 +144,10 @@ public class PartitionServiceImpl implements IPartitionService {
   @Transactional
   @Override
   public boolean deletePartition(String partitionId) {
-    this.partitionPropertyEntityRepository.findByPartitionId(partitionId)
-        .orElseThrow(
-            () -> new AppException(HttpStatus.SC_NOT_FOUND, UNKNOWN_ERROR_REASON,
-                "An attempt to delete not existing partition."));
+    if (!this.partitionPropertyEntityRepository.findByPartitionId(partitionId).isPresent()) {
+      throw new AppException(HttpStatus.SC_NOT_FOUND, UNKNOWN_ERROR_REASON,
+          "An attempt to delete not existing partition.");
+    }
     this.partitionPropertyEntityRepository.deleteByPartitionId(partitionId);
     return true;
   }
